gcloud_functions/multi-source-agent/requirements.txt
functions-framework==3.*
google-genai==1.53.0
httpx==0.28.1